fix: resolve clippy diagnostics across workspace
Fix 29 clippy/rustc diagnostics reported by the IDE:
- approx_constant (error): replace 3.14/3.14159 literals with
  std::f64::consts::PI and FRAC_PI_2
- unnecessary_cast: remove redundant `as u32` casts on u32 expressions
- needless_range_loop: replace index-based loops with iter().enumerate()
- items_after_test_module: move color_magnitude() before test module
  in analysis.rs and remove the duplicate definition
- unused_variables: prefix unused r/b/y/y2 with underscore
- manual_range_contains: use (5..=6).contains(&val)
- len_zero: use !is_empty() instead of len() >= 1

+++ crates/leptonica-core/src/pix/clip.rs
@@ -1459,7 +1459,7 @@ mod tests {
             let mut pm = base.try_into_mut().unwrap();
             for y in 0..10u32 {
                 for x in 0..10u32 {
-                    pm.set_pixel_unchecked(x, y, (x + y * 10) as u32);
+                    pm.set_pixel_unchecked(x, y, x + y * 10);
                 }
             }
             Pix::from(pm)
